Polish folder source control cards
