+++ src/renderer/src/i18n/locales/en.json
@@ -11775,7 +11775,8 @@
           "cached": "{{value0}} cached",
           "pending": "Status pending",
           "changeOne": "1 change",
-          "changeMany": "{{value0}} changes"
+          "changeMany": "{{value0}} changes",
+          "jumpToWorktree": "Jump to worktree"
         }
       },
       "link": {
